패널 우량 버그 수정

+++ application/app/panel/panel_mooju.c
@@ -120,11 +120,11 @@ framemk[cnt++] 		= 17;																	// Length
 
 sprintf(&framemk[cnt],"%5.1f", (float)p_kma->wind_speed_avg.data / 10.0);						// 풍속
 cnt					+= 5;
-sprintf(&framemk[cnt],"%4d", g_rainfall.yearly/10.0);					// 연간 누계 강우량
+sprintf(&framemk[cnt],"%4d", (int)(g_rainfall.yearly/10.0));					// 연간 누계 강우량
 cnt					+= 4;
-sprintf(&framemk[cnt], "%4d", g_rainfall.today / 10.0); // 금일  강우량
+sprintf(&framemk[cnt], "%4d", (int)(g_rainfall.today / 10.0)); // 금일  강우량
 cnt					+= 4;
-sprintf(&framemk[cnt], "%4d", g_rainfall.yesterday / 10.0); // 전일  강우량
+sprintf(&framemk[cnt], "%4d", (int)(g_rainfall.yesterday / 10.0)); // 전일  강우량
 cnt					+= 4;
 
 framemk[cnt++]		= (char)make_sum((uint8_t*)&framemk[1], framemk[4]+4);
